lets make this a component
